+++ Release_Notes.md
@@ -13,6 +13,7 @@
 
 ### Bug Fixes
 - Fixed NG0956: The configured tracking expression (track by identity) in file-list.component.html
+- Fixed Pager not displaying the correct number of pages
 
 ## Net_Core Version 5.0.0.x (2024-08-31)